# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: JornadaMS API
  version: 1.0.0-mvp1
  description: API REST privada do MVP-1 do JornadaMS.
servers:
  - url: http://localhost:8000
    description: Desenvolvimento local
...
security:
  - BearerAuth: []
  - CookieAuth: []
paths:
...
  /api/v1/work-schedules:
    get:
      tags: [Schedules]
      operationId: listWorkSchedules
      x-mvp: MVP-1
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - in: query
          name: status
          schema:
            $ref: '#/components/schemas/ScheduleStatus'
      responses:
        '200':
          description: Lista de jornadas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkSchedulePage'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags: [Schedules]
      operationId: createWorkSchedule
      x-mvp: MVP-1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkScheduleCreate'
      responses:
        '201':
          description: Jornada criada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkSchedule'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/ValidationError'
  /api/v1/work-schedules/{schedule_id}:
    parameters:
      - $ref: '#/components/parameters/ScheduleId'
    get:
      tags: [Schedules]
      operationId: getWorkSchedule
      x-mvp: MVP-1
      responses:
        '200':
          description: Jornada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkSchedule'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Schedules]
      operationId: updateWorkSchedule
      x-mvp: MVP-1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkScheduleUpdate'
      responses:
        '200':
          description: Jornada atualizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkSchedule'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
  /api/v1/employees/{employee_id}/schedules:
    post:
      tags: [Schedules]
      operationId: assignEmployeeSchedule
      x-mvp: MVP-1
      parameters:
        - $ref: '#/components/parameters/EmployeeId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmployeeScheduleCreate'
      responses:
        '201':
          description: Jornada vinculada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeSchedule'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    CookieAuth:
      type: apiKey
      in: cookie
      name: jornada_access
  parameters:
    ScheduleId:
      in: path
      name: schedule_id
      required: true
      schema:
        type: string
        format: uuid
    EmployeeId:
      in: path
      name: employee_id
      required: true
      schema:
        type: string
        format: uuid
...
    Page:
      in: query
      name: page
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      in: query
      name: page_size
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 50
...
  responses:
    Unauthorized:
      description: Não autenticado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Sem permissão
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Recurso não encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Conflito de estado ou idempotência
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ValidationError:
      description: Dados semanticamente inválidos
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    ScheduleStatus:
      type: string
      enum: [ACTIVE, INACTIVE]
...
    WorkScheduleCreate:
      type: object
      required: [name, start_time, end_time]
      properties:
        name:
          type: string
        start_time:
          type: string
          format: time
        end_time:
          type: string
          format: time
        same_day_only:
          type: boolean
          default: true
        tolerance_minutes:
          type: integer
          minimum: 0
          default: 0
    WorkScheduleUpdate:
      type: object
      description: Campos opcionais para atualização parcial da jornada.
      properties:
        name:
          type: string
          minLength: 1
        start_time:
          type: string
          format: time
        end_time:
          type: string
          format: time
        same_day_only:
          type: boolean
        tolerance_minutes:
          type: integer
          minimum: 0
    WorkSchedule:
      allOf:
        - $ref: '#/components/schemas/WorkScheduleCreate'
        - type: object
          required: [id, status]
          properties:
            id:
              type: string
              format: uuid
            status:
              $ref: '#/components/schemas/ScheduleStatus'
    EmployeeScheduleCreate:
      type: object
      required: [schedule_id, starts_on]
      properties:
        schedule_id:
          type: string
          format: uuid
        starts_on:
          type: string
          format: date
        ends_on:
          type: [string, 'null']
          format: date
    EmployeeSchedule:
      allOf:
        - $ref: '#/components/schemas/EmployeeScheduleCreate'
        - type: object
          required: [employee_id]
          properties:
            employee_id:
              type: string
              format: uuid
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, correlation_id]
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: array
              items: {}
            correlation_id:
              type: string
...
    WorkSchedulePage:
      type: object
      required: [items, page, page_size, total_items, total_pages]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/WorkSchedule'
        page:
          type: integer
        page_size:
          type: integer
        total_items:
          type: integer
        total_pages:
          type: integer
...